update navbar & refactor api calls
