+++ locales/en/common.json
@@ -151,6 +151,7 @@
     "message": "This feature is not yet available. Subscribe to our newsletter and get notified when it's ready.",
     "success": "Thank you!"
   },
+  "new": "New!",
   "year-filter-info": "Author's death date",
   "regions-filter-info": "Locations where the author has visited, been born, or died",
   "text-info": "Text Info",
