+++ src/app/[locale]/collections/[slug]/collection-books/genres-filter.tsx
@@ -4,9 +4,9 @@ import { useMemo, useState } from "react";
 import FilterContainer from "@/components/search-results/filter-container";
 import { Button } from "@/components/ui/button";
 import { Skeleton } from "@/components/ui/skeleton";
+import { findAllGenresWithBooksCount } from "@/lib/api/genres";
 import { usePathLocale } from "@/lib/locale/utils";
 import { usePathname } from "@/navigation";
-import { findAllGenresWithBooksCount } from "@/server/services/genres";
 import { useQuery } from "@tanstack/react-query";
 import Fuse from "fuse.js";
 import { useFormatter, useTranslations } from "next-intl";
