+++ src/app/manifest.ts
@@ -8,12 +8,12 @@ export default function manifest(): MetadataRoute.Manifest {
     description: "Usul - The Research Tool for Islamic Texts",
     icons: [
       {
-        src: "/android-chrome-192x192.png",
+        src: "/favicons/android-chrome-192x192.png",
         sizes: "192x192",
         type: "image/png",
       },
       {
-        src: "/android-chrome-512x512.png",
+        src: "/favicons/android-chrome-512x512.png",
         sizes: "512x512",
         type: "image/png",
       },
